fix default empty values not present in the secret
I'm didn't dig deep enough to find out the root cause of this. The issue was that on helm install, the `AWS_SESSION_TOKEN` and `aws-assume-role` would be correctly present in the secret with an empty string as the value, however after a helm update with the same values, those keys would be completely removed from the secret. Quoting it fixes the issue. My guess is that something is being passed as a nil probably from the b64enc and either helm or k8s is not including the field